adds rating logic, styling, buttons

+++ src/views/ProfessorShow.vue
@@ -4,12 +4,17 @@
     
     <div class="professor">
       <div class="professor-info">
-        <h1 id="score">{{ professorScore }}</h1>
-        <p>{{ professor.title }}</p>
+        <div class="rating">
+          <h1 id="score">{{ professorScore }}</h1>
+          <i class="material-icons">star_rate</i><i class="material-icons">star_rate</i><i class="material-icons">star_rate</i><i class="material-icons">star_rate</i>
+
+          <p>{{ professor.title }}</p>
+        </div>
         <h1>{{ professor.name }}</h1>
         <h3>{{ professor.school }}</h3>
         <p>{{ professor.department }}</p>
         <p>{{ professor.subject }}</p>
+        <div class="image"></div>
       </div>
     </div>
 
@@ -17,15 +22,22 @@
 
     <!-- Review Index -->
     <div class="reviews">
-      <h2>REVIEWS</h2>
-      <button v-on:click="addReview()">Add Review</button>
+      <div class="row reviews-bar">
+        <h2>REVIEWS</h2><br>
+        <div class=buttonholder>
+          <button class="btn-tertiary" style="color:#505050;">Sort<i class="material-icons">arrow_drop_down</i></button>
+          <button class="btn-tertiary" style="color:#505050;">Filter<i class="material-icons">arrow_drop_down</i></button>
+          <button v-on:click="addReview()"><i class="material-icons" style="font-size:1.2em;font-weight:900;">add</i>Add Review</button>
+        </div>
+      </div>
       <div class="review-index" v-for="review in reviews">
         <h4>{{ review.author }} - {{ review.date | moment('MMM Do, YYYY')}}</h4>
         <p>Score: {{ review.score }}</p>
-        <p>{{ review.text }}</p>
-        <button v-on:click="editReview(review)">Edit Review</button>
-        <button v-on:click="deleteReview(review.id)">Delete Review</button>
+        <p>{{ review.text }}</p> 
+        <button class="btn-secondary" v-on:click="editReview(review)"><i class="material-icons" style="font-size:1.2em;">edit</i>Edit Review</button>
+        <button class="btn-tertiary" v-on:click="deleteReview(review.id)">Delete Review</button>
       </div>
+      <h2 v-if="reviews.length===0" style="text-align: center;font-weight:300;"><br>No Reviews Yet!</h2>
     </div>
 
     <!-- Create Review -->
@@ -34,7 +46,7 @@
         Author:<input type ="text" v-model="newReviewAuthor"></input>
         Date:<input type ="date" v-model="newReviewDate"></input>
         Text:<input type ="text" v-model="newReviewText"></input>
-        Score:<input type ="number" step="0.1" v-model="newReviewScore"></input>
+        Score:<input type ="number" step="0.5" v-model="newReviewScore"></input>
         Professor_Id:<input type ="number" v-model="newReviewProfessorId"></input>
         <button v-on:click="createReview()">Create Review</button>
       </form>
@@ -44,8 +56,8 @@
     <dialog id="review-edit">
       <form>
         Author:<input type ="text" v-model="currentReview.author"></input><br>
-        Text:<input type ="text" v-model="currentReview.text"></input><br>
-        Score:<input type ="number" step="0.1" v-model="currentReview.score"></input><br>
+        Text:<input type ="text" style="height: 200px;width: 200px;" v-model="currentReview.text"></input><br>
+        Score:<input type ="number" step="0.5" v-model="currentReview.score"></input><br>
         Professor_Id:<input type ="number" v-model="currentReview.professor_id"></input><br>
         <button v-on:click="updateReview()">Update Review</button><br>
       </form>
@@ -91,6 +103,8 @@
   border-radius: 7px;
   vertical-align: top;
   letter-spacing: 0.18em;
+  box-shadow: 6px 14px 29px 0px rgba(41, 41, 41, 0.08);
+  /* line-height: 1.33em; */
 }
 
 .professor #score {
@@ -101,19 +115,106 @@
   text-align: center;
 }
 
+.professor i {
+  color: #dc9731;
+}
+
+.professor h1 {
+  font-weight: 900 !important;
+}
+
+.professor .image {
+  height: 64px;
+  width: 64px;
+  background-color: gray;
+  margin: 3rem auto 1rem auto;
+}
+
+.professor .rating {
+  text-align: center;
+  margin-bottom: 6rem;
+}
+
 .reviews {
   display: inline-block;
   width: 45%;
   text-align: left;
-  padding: 2.5em;
+  padding: 1.5em 6.5em 2.5em 3.5em;
   margin: 2.5em;
-
   background-color: #ffffff;
   border-radius: 7px;
+  box-shadow: 8px 20px 33px 0px rgba(41, 41, 41, 0.08);
 }
 
 .reviews h4 {
   text-transform: uppercase;
+  margin-top: 4em;
+  margin-bottom: 0;
+  line-height: 0.1em;
+}
+
+.reviews-bar {
+  width: 100%;
+}
+
+.reviews-bar h2 {
+  float: left;
+  width: 100px;
+}
+
+.reviews p {
+  line-height: 1.6em;
+}
+
+.review-index button {
+  margin-right: 1.3em;
+}
+
+.buttonholder {
+  width: 450px;
+  text-align: right;
+  display: block;
+  margin: 0 0 0 auto;
+}
+
+.buttonholder button {
+  align-content: top;
+  margin-left: 1.1rem;
+}
+
+button {
+  background-color: #73a2fd;
+  font-family: "Lato", sans-serif;
+  font-weight: 700;
+  border-radius: 3px;
+  border-style: none;
+  padding: 0.6rem;
+  color: white;
+  text-transform: uppercase;
+  cursor: pointer;
+}
+
+i {
+  padding-right: 0.3em;
+  vertical-align: bottom;
+}
+
+.btn-secondary {
+  background-color: #cccccc;
+}
+
+.btn-tertiary {
+  background-color: rgba(0, 0, 0, 0);
+  color: #999999;
+}
+
+dialog {
+  height: 650px;
+  width: 400px;
+}
+
+dialog form {
+  text-align: left;
 }
 </style>
 <script>
@@ -149,6 +250,7 @@ export default {
       this.calculateScore();
     });
   },
+  mounted: function () {},
   methods: {
     addReview: function () {
       document.querySelector("#reviewnew").showModal();
@@ -170,9 +272,11 @@ export default {
         });
     },
     deleteReview: function (id) {
-      axios.delete(`/reviews/${id}`).then((response) => {
-        console.log("Review Deleted");
-      });
+      if (confirm("Are you sure you want to delete this review?")) {
+        axios.delete(`/reviews/${id}`).then((response) => {
+          console.log("Review Deleted");
+        });
+      }
     },
     // getProfessorName: function (professorId) {
     //   var index = this.professors.find((x) => x.id === professorId);
@@ -229,10 +333,9 @@ export default {
       this.reviews.forEach(function (index) {
         sumTotal += index.score;
       });
-      console.log(sumTotal);
       var result = sumTotal / this.reviews.length;
       console.log("Rating:", result);
-      this.professorScore = result;
+      this.professorScore = result.toFixed(1);
     },
   },
 };
